requirements.txt
django>=3.2.4
djangorestframework>=3.12.4
django-cors-headers>=3.7.0
django-crum>=0.7.9
requests_html